requirements.txt
beautifulsoup4==4.4.1
Django==1.10.1
django-ajax-selects==1.5.2
django-crispy-forms==1.6.1
django-widget-tweaks==1.4.1
fuzzywuzzy==0.14.0
pycountry==16.11.8
titlecase==0.8.1
urllib3==1.13.1
requests==2.11.1
simplejson==3.8.2

